refactor: implement modular shell architecture with config and preview panels
